Restructured OS specific /etc/hosts related files
