+++ pkg/kyma/cmd/provision/minikube/cfg_others.go
@@ -1,0 +1,8 @@
+// +build !windows
+// +build !darwin
+
+package minikube
+
+const hostsFile = "/etc/hosts"
+
+const defaultVMDriver = vmDriverNone
