+++ pkg/kyma/cmd/provision/minikube/cfg_windows.go
@@ -1,0 +1,7 @@
+// +build windows
+
+package minikube
+
+const hostsFile = "C:\\Windows\\system32\\drivers\\etc\\hosts"
+
+const defaultVMDriver = vmDriverVirtualBox
