+++ pkg/kyma/cmd/provision/minikube/hosts_windows.go
@@ -8,10 +8,6 @@ import (
 	"github.com/kyma-project/cli/internal/step"
 )
 
-const hostsFile = "C:\\Windows\\system32\\drivers\\etc\\hosts"
-
-const defaultVMDriver = vmDriverVirtualBox
-
 func addDevDomainsToEtcHostsOSSpecific(o *MinikubeOptions, s step.Step, hostAlias string) error {
 
 	s.LogErrorf("Please add these lines to your " + hostsFile + " file:")
